requirements.txt
python-dateutil~=2.9.0.post0
matplotlib~=3.8.3
pytz~=2024.1
numpy~=1.26.4